Update ca.json change "Escriptura" with "Redacció". Adding translators information.
"Escriptura" is right but sounds kind of forced.

As in the Spanish translation, we asked catalan native Editorial fellows and they felt "Redacció" fits better with the task. Notice the explanation of the task talks about "redacció", so is better keeping the match between the title and the description (as we do in English).

Translators info was updated.